02.03.2021 Correction

+++ src/main/java/com/springboot/crud_security/entity/Role.java
@@ -1,7 +1,6 @@
 package com.springboot.crud_security.entity;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.fasterxml.jackson.annotation.JsonValue;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -22,7 +21,7 @@ public class Role implements GrantedAuthority {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private int id;
+    private Long id;
 
 //    @Enumerated(value = EnumType.STRING)
 //    @Column(name = "name")
